Added a Webserver and OTA, removed OLED
Added a Webserver and OTA functions. I did not need the OLED as my device is outside my house near the inverter that has it's own display so I removed it to save memory and CPU cycles. We can now see if the device is online via the webserver and best of all update it without leaving my desk! 
http://sofar2mqtt.local/ - Basic page with the some basic info values can be seen in the json output.
http://sofar2mqtt.local/jsonOut - will provide a Json output for those who do not use MQTT
http://sofar2mqtt.local/reboot - Yep it reboots the device.